configured redux, added modalSlice
