+++ src/redux/store.ts
@@ -3,12 +3,11 @@
 import { configureStore } from "@reduxjs/toolkit";
 import modalReducer from "./modal-slice";
 
-export const store = () =>
-    configureStore({
-        reducer: {
-            modal: modalReducer,
-        },
-    });
+export const store = configureStore({
+    reducer: {
+        modal: modalReducer,
+    },
+});
 
 export type RootState = ReturnType<typeof store.getState>;
 export type AppDispatch = typeof store.dispatch;
