--- src/store/modal-slice.ts
@@ -1,11 +1,0 @@
-import { createSlice } from "@reduxjs/toolkit";
-
-const initialState = {} as any;
-
-const modalSlice = createSlice({
-    name: "modal",
-    initialState,
-    reducers: {},
-});
-
-export default modalSlice.reducer;
