--- src/store/provider.tsx
@@ -1,8 +1,0 @@
-"use client";
-
-import { Provider } from "react-redux";
-import { store } from "./store";
-
-export const Providers = ({ children }) => {
-    return <Provider store={store()}>{children}</Provider>;
-};
